requirements.txt
fastapi==0.115.12
tortoise-orm==0.25.0
uvicorn==0.34.2
aiomysql==0.2.0
pyjwt==2.10.1
bcrypt==4.3.0
passlib==1.7.4
python-multipart==0.0.20